stdlib: Clean up init/fini handling
Always make __call_exitprocs "weak" (which used to be the "lite-exit"
option). This avoids pulling in all of that code when unused.

Make newlib-initfini false by default; I've never seen those functions
used. Process preinit and init functions in the same loop when
newlib-initfini is false.
