+++ COPYING.picolibc
@@ -1120,6 +1120,7 @@ Files: newlib/empty.c
  newlib/libc/stdlib/pico-atexit.c
  newlib/libc/stdlib/pico-cxa-atexit.c
  newlib/libc/stdlib/pico-exit.c
+ newlib/libc/stdlib/pico-exitprocs.c
  newlib/libc/stdlib/pico-onexit.c
  newlib/libc/stdlib/pico-onexit.h
  newlib/libc/tinystdio/atold_engine.c
