+++ meson.build
@@ -182,7 +182,6 @@ endif
 newlib_atexit_dynamic_alloc = get_option('newlib-atexit-dynamic-alloc')
 newlib_nano_malloc = get_option('newlib-nano-malloc')
 nano_malloc_clear_freed = get_option('nano-malloc-clear-freed')
-lite_exit = get_option('lite-exit')
 
 newlib_elix_level = get_option('newlib-elix-level')
 c_args = core_c_args
@@ -1405,7 +1404,7 @@ conf_data.set('_MB_EXTENDED_CHARSETS_JIS', mb_jis_charsets)
 conf_data.set('__SINGLE_THREAD__', get_option('newlib-multithread') == false)
 conf_data.set('_ELIX_LEVEL', newlib_elix_level)
 conf_data.set('_ATEXIT_DYNAMIC_ALLOC', newlib_atexit_dynamic_alloc)
-conf_data.set('_REENT_GLOBAL_ATEXIT', get_option('newlib-global-atexit'))
+conf_data.set('_GLOBAL_ATEXIT', get_option('newlib-global-atexit'))
 conf_data.set('_FVWRITE_IN_STREAMIO', newlib_fvwrite_in_streamio)
 conf_data.set('_FSEEK_OPTIMIZATION', newlib_fseek_optimization)
 conf_data.set('_WIDE_ORIENT', newlib_wide_orient)
@@ -1413,7 +1412,6 @@ conf_data.set('_HAVE_FCNTL', newlib_have_fcntl)
 conf_data.set('_NANO_MALLOC', newlib_nano_malloc)
 conf_data.set('_NANO_MALLOC_CLEAR_FREED', nano_malloc_clear_freed and newlib_nano_malloc)
 conf_data.set('_UNBUF_STREAM_OPT', get_option('newlib-unbuf-stream-opt'))
-conf_data.set('_LITE_EXIT', lite_exit)
 conf_data.set('_PICO_EXIT', picoexit)
 conf_data.set('_NANO_FORMATTED_IO', newlib_nano_formatted_io)
 conf_data.set('_RETARGETABLE_LOCKING', get_option('newlib-retargetable-locking'))
