+++ newlib/libc/stdlib/__atexit.c
@@ -43,11 +43,9 @@
 void * malloc(size_t) _ATTRIBUTE((__weak__));
 #endif
 
-#ifdef _LITE_EXIT
 /* As __call_exitprocs is weak reference in lite exit, make a
    non-weak reference to it here.  */
 const void * __atexit_dummy = &__call_exitprocs;
-#endif
 
 NEWLIB_THREAD_LOCAL_ATEXIT struct _atexit _atexit0;
 NEWLIB_THREAD_LOCAL_ATEXIT struct _atexit *_atexit;
