+++ newlib/libc/stdlib/pico-onexit.c
@@ -35,79 +35,11 @@
 
 #define _DEFAULT_SOURCE
 #include <stdlib.h>
-#include <string.h>
-#include <sys/lock.h>
-#include <limits.h>
 #include "pico-onexit.h"
-#include "atexit.h"
-
-struct on_exit {
-    union on_exit_func  func;
-    void                *arg;
-    int                 kind;
-};
-
-static struct on_exit on_exits[ATEXIT_MAX];
-
-int
-_on_exit(int kind, union on_exit_func func, void *arg)
-{
-	int	ret = -1;
-	int	o;
-	__LIBC_LOCK();
-	for (o = 0; o < ATEXIT_MAX; o++) {
-		if (on_exits[o].kind == PICO_ONEXIT_EMPTY) {
-			on_exits[o].func = func;
-			on_exits[o].arg = arg;
-                        on_exits[o].kind = kind;
-			ret = 0;
-			break;
-		}
-	}
-	__LIBC_UNLOCK();
-	return ret;
-}
 
 int
 on_exit (void (*func)(int, void *),void *arg)
 {
     union on_exit_func func_u = { .on_exit = func };
     return _on_exit(PICO_ONEXIT_ONEXIT, func_u, arg);
-}
-
-void
-__call_exitprocs(int code, void *param)
-{
-        (void) param;
-	for (;;) {
-		int	                i;
-                union on_exit_func      func = {0};
-                int                     kind = PICO_ONEXIT_EMPTY;
-		void	                *arg = 0;
-
-		__LIBC_LOCK();
-		for (i = ATEXIT_MAX - 1; i >= 0; i--) {
-                        kind = on_exits[i].kind;
-			if (kind != PICO_ONEXIT_EMPTY) {
-                                func = on_exits[i].func;
-                                arg = on_exits[i].arg;
-                                memset(&on_exits[i], '\0', sizeof(struct on_exit));
-				break;
-			}
-		}
-		__LIBC_UNLOCK();
-                switch (kind) {
-                case PICO_ONEXIT_EMPTY:
-                        return;
-                case PICO_ONEXIT_ONEXIT:
-                        func.on_exit(code, arg);
-                        break;
-                case PICO_ONEXIT_ATEXIT:
-                        func.atexit();
-                        break;
-                case PICO_ONEXIT_CXA_ATEXIT:
-                        func.cxa_atexit(arg);
-                        break;
-                }
-	}
 }
