+++ newlib/libc/stdlib/pico-onexit.h
@@ -33,10 +33,15 @@
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#define PICO_ONEXIT_EMPTY       0
-#define PICO_ONEXIT_ONEXIT      1
-#define PICO_ONEXIT_ATEXIT      2
-#define PICO_ONEXIT_CXA_ATEXIT  3
+#ifndef _PICO_ONEXIT_H_
+#define _PICO_ONEXIT_H_
+
+enum pico_onexit_kind {
+    PICO_ONEXIT_EMPTY,
+    PICO_ONEXIT_ONEXIT,
+    PICO_ONEXIT_ATEXIT,
+    PICO_ONEXIT_CXA_ATEXIT
+};
 
 union on_exit_func {
     void        (*on_exit)(int, void *);
@@ -45,4 +50,9 @@ union on_exit_func {
 };
 
 int
-_on_exit(int kind, union on_exit_func func, void *arg);
+_on_exit(enum pico_onexit_kind kind, union on_exit_func func, void *arg);
+
+void
+__call_exitprocs (int, void *);
+
+#endif /* _PICO_ONEXIT_H_ */
