+++ picolibc.ld.in
@@ -81,6 +81,7 @@ SECTIONS
 
 		/* lists of constructors and destructors */
 		PROVIDE_HIDDEN ( @PREFIX@__preinit_array_start = . );
+		PROVIDE_HIDDEN ( @PREFIX@__bothinit_array_start = . );
 		KEEP (*(.preinit_array))
 		PROVIDE_HIDDEN ( @PREFIX@__preinit_array_end = . );
 
@@ -88,6 +89,7 @@ SECTIONS
 		KEEP (*(SORT_BY_INIT_PRIORITY(.init_array.*) SORT_BY_INIT_PRIORITY(.ctors.*)))
 		KEEP (*(.init_array .ctors))
 		PROVIDE_HIDDEN ( @PREFIX@__init_array_end = . );
+		PROVIDE_HIDDEN ( @PREFIX@__bothinit_array_end = . );
 
 		PROVIDE_HIDDEN ( @PREFIX@__fini_array_start = . );
 		KEEP (*(SORT_BY_INIT_PRIORITY(.fini_array.*) SORT_BY_INIT_PRIORITY(.dtors.*)))
